Add slice frame functionality
Before this, the aggregate level of charging depended on the number of slices with something in them. This ignored the fact that many perfectly valid slices have no events in them. By adding a slice frame, I identify all of the possible slices that could be sampled from. Then I sample from empty and non-empty slices using P(slice) = P(slice | is nonempty)P(is nonempty) + P(slice | is empty)P(is empty)