Switch sorting in profile creation of slice_vehicle_windows to be based on vehicle and UTC time only
Used to be vehicle, local time, and the profile column, but this is no longer necessary. Duplicate dropping just introduced ensures that we don't have multiple events in the same time which could get confused in their order. The switch to UTC time for sorting is essential for dealing with DST